moved visualizer out into its own module which included removing the CLIOutput dependency of the class
