+++ vminfo_parser/vminfo_parser.py
@@ -18,6 +18,7 @@ import pandas as pd
 
 from . import const
 from .config import Config
+from .visualizer import Visualizer
 
 A = t.TypeVar("Analyzer", bound="Analyzer")
 LOGGER = logging.getLogger(__name__)
@@ -346,163 +347,6 @@ class CLIOutput:
             else:
                 self.writeline("No data available for the specified resource.")
             self.writeline("")
-
-
-class Visualizer:
-    def __init__(self: t.Self) -> None:
-        self.cli_output = CLIOutput()
-
-    @classmethod
-    def visualize_disk_space(
-        cls: t.Self,
-        disk_space_ranges: t.Optional[list[tuple[int, int]]],
-        dataFrame: pd.DataFrame,
-        column_headers: t.Optional[dict[str, str]],
-    ) -> None:
-
-        # Count the number of VMs in each disk space range
-        range_counts = {range_: 0 for range_ in disk_space_ranges}
-        for lower, upper in disk_space_ranges:
-            mask = (dataFrame[column_headers["vmDisk"]] >= lower) & (dataFrame[column_headers["vmDisk"]] <= upper)
-            count = dataFrame.loc[mask].shape[0]
-            range_counts[(lower, upper)] = count
-
-        # Sort the counts and plot them as a horizontal bar chart
-        sorted_dict = dict(sorted(range_counts.items(), key=lambda x: x[1]))
-
-        if dataFrame.empty:
-            LOGGER.warning("No data to plot")
-            return
-
-        # Create a subplot for plotting
-        fig, ax = plt.subplots()
-
-        # Plot the sorted counts as a horizontal bar chart
-        for range_, count in sorted_dict.items():
-            ax.barh(f"{range_[0]}-{range_[1]} GB", count)
-
-        # Set titles and labels for the plot
-        ax.set_ylabel("Disk Space Range")
-        ax.set_xlabel("Number of VMs")
-        ax.xaxis.set_major_formatter(ticker.ScalarFormatter())
-
-        ax.set_title("Hard Drive Space Breakdown for Organization")
-
-        ax.set_xlim(right=max(range_counts.values()) + 1.5)
-
-        # Display the plot
-        plt.show(block=True)
-        plt.close()
-
-    def visualize_disk_space_distribution(
-        self: t.Self,
-        range_counts_by_environment: pd.DataFrame,
-        environment_filter: str,
-        os_filter: t.Optional[str] = None,
-    ) -> None:
-        range_counts_by_environment.plot(kind="bar", stacked=False, figsize=(12, 8), rot=45)
-
-        plt.xlabel("Disk Space Range")
-        plt.ylabel("Number of VMs")
-        plt.title(f'VM Disk Size Ranges Sorted by Environment {f"for {os_filter}" if os_filter else ""}')
-
-        plt.show(block=True)
-        plt.close()
-
-    def visualize_os_distribution(
-        self: t.Self,
-        counts: pd.Series,
-        os_names: list[str],
-        dataFrame: pd.DataFrame,
-        environment_filter: t.Optional[str] = None,
-        min_count: int = 500,
-    ) -> None:
-        # Define specific colors for identified OS names
-        # Generate random colors for bars not in SUPPORTED_OS_COLORS
-        random_colors = cm.rainbow(np.linspace(0, 1, len(counts)))
-        colors = [const.SUPPORTED_OS_COLORS.get(os, random_colors[i]) for i, os in enumerate(os_names)]
-
-        # Plot the counts as a horizontal bar chart with specified and random colors
-        # ax = counts.plot(kind="barh", rot=45, color=colors)
-        counts.plot(kind="barh", rot=45, color=colors)
-
-        if dataFrame.empty:
-            LOGGER.warning("No data to plot")
-            return
-
-        # Set titles and labels for the plot
-        plt.title(f"OS Counts by Environment Type (>= {min_count})")
-        plt.xlabel("Count")
-        plt.ylabel("Operating Systems")
-
-        # Display the plot
-        plt.show(block=True)
-        plt.close()
-
-    def visualize_unsupported_os_distribution(self: t.Self, unsupported_counts: pd.Series) -> None:
-
-        random_colors = cm.rainbow(np.linspace(0, 1, len(unsupported_counts)))
-        plt.pie(
-            unsupported_counts,
-            labels=unsupported_counts.index,
-            colors=random_colors,
-            autopct="%1.1f%%",
-        )
-        plt.title("Unsupported Operating System Distribution")
-
-        plt.show(block=True)
-        plt.close()
-
-    def visualize_supported_os_distribution(
-        self: t.Self,
-        counts: pd.Series,
-        environment_filter: t.Optional[str] = None,
-    ) -> None:
-        colors = [const.SUPPORTED_OS_COLORS[os] for os in counts.index]
-
-        if environment_filter and environment_filter != "both":
-            counts.plot(kind="barh", rot=45, color=colors)
-        else:
-            counts.plot(kind="barh", rot=45)
-
-        if environment_filter not in ["prod", "non-prod"]:
-            plt.title("Supported Operating Systems For All Environments")
-        else:
-            plt.title(f"Supported Operating Systems for {environment_filter.title()}")
-
-        plt.ylabel("Operating Systems")
-        plt.xlabel("Count")
-        plt.xscale("log")
-        plt.gca().xaxis.set_major_formatter(ticker.ScalarFormatter())
-
-        if environment_filter != "both":
-            plt.xticks(
-                [
-                    counts.iloc[0] - (counts.iloc[0] % 100),
-                    counts.iloc[len(counts) // 2] - (counts.iloc[len(counts) // 2] % 100),
-                    counts.iloc[-1],
-                ]
-            )
-
-        plt.show(block=True)
-        plt.close()
-
-    def visualize_os_version_distribution(self: t.Self, os_name: str, dataFrame: pd.DataFrame, config: Config) -> None:
-        counts = self.cli_output.generate_os_version_distribution(dataFrame, os_name, config.minimum_count)
-
-        if not counts.empty:
-            ax = counts.plot(kind="barh", rot=45)
-            plt.title(f"Distribution of {os_name}")
-            plt.ylabel("OS Version")
-            plt.xlabel("Count")
-
-            plt.xticks(rotation=0)
-            ax.set_yticklabels(counts["OS Version"])
-
-            plt.show(block=True)
-            plt.close()
-
-        self.cli_output.format_dataframe_output(counts, os_name=os_name)
 
 
 class Analyzer:
@@ -851,6 +695,7 @@ def main(*args: t.Optional[str]) -> None:  # noqa: C901
 
     analyzer = Analyzer(vm_data, config, column_headers=vm_data.column_headers)
     visualizer = Visualizer()
+    cli_output = CLIOutput()
 
     # Load environments from prod-env-labels if provided
     environments = []
@@ -966,7 +811,11 @@ def main(*args: t.Optional[str]) -> None:  # noqa: C901
         if config.generate_graphs:
             for os_name in vm_data.df["OS Name"].unique():
                 if os_name is not None and not pd.isna(os_name) and os_name != "":
-                    visualizer.visualize_os_version_distribution(os_name, vm_data.df, config)
+                    counts_dataframe = cli_output.generate_os_version_distribution(
+                        vm_data.df, os_name, config.minimum_count
+                    )
+                    cli_output.format_dataframe_output(counts_dataframe, os_name=os_name)
+                    visualizer.visualize_os_version_distribution(os_name, vm_data.df, config, counts_dataframe)
 
     if config.get_supported_os:
         if config.prod_env_labels and config.sort_by_env:
@@ -991,7 +840,7 @@ def main(*args: t.Optional[str]) -> None:  # noqa: C901
 
     # close clioutput
     analyzer.cli_output.close()
-    visualizer.cli_output.close()
+    cli_output.close()
 
 
 if __name__ == "__main__":
